sub_806C4D4

+++ include/pokemon_mid.h
@@ -6,7 +6,7 @@
 bool8 ComparePokemonNames(s16, s16);
 void CopyCyanMonsterNametoBuffer(u8 *buffer, s16 index);
 void PrintPokeNameToBuffer(u8 *buffer, PokemonStruct1 *pokemon);
-void GetPokemonLevelData(LevelData* a1, s16 _id, s32 level);
+void GetPokemonLevelData(LevelData* a1, s32 _id, s32 level);
 const u8* DecompressMoveID(const u8* a1, u16* moveID);
 
 bool8 sub_808DA44(s32, u32);
